move gizmos to their own files, fix typo
